+++ src/main/java/org/example/mollyapi/payment/service/impl/PaymentServiceImpl.java
@@ -1,6 +1,7 @@
 package org.example.mollyapi.payment.service.impl;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,9 +12,12 @@ import org.example.mollyapi.common.exception.CustomException;
 import org.example.mollyapi.common.exception.error.impl.PaymentError;
 import org.example.mollyapi.common.exception.error.impl.UserError;
 import org.example.mollyapi.delivery.dto.DeliveryReqDto;
+import org.example.mollyapi.delivery.entity.Delivery;
+import org.example.mollyapi.delivery.repository.DeliveryRepository;
 import org.example.mollyapi.order.entity.Order;
+import org.example.mollyapi.order.entity.OrderDetail;
 import org.example.mollyapi.order.repository.OrderRepository;
-import org.example.mollyapi.order.service.OrderService;
+import org.example.mollyapi.order.type.OrderStatus;
 import org.example.mollyapi.payment.dto.request.PaymentCancelReqDto;
 import org.example.mollyapi.payment.dto.request.TossCancelReqDto;
 import org.example.mollyapi.payment.dto.response.PaymentInfoResDto;
@@ -26,6 +30,8 @@ import org.example.mollyapi.payment.repository.PaymentRepository;
 import org.example.mollyapi.payment.service.PaymentService;
 import org.example.mollyapi.payment.util.MapperUtil;
 import org.example.mollyapi.payment.util.PaymentWebClientUtil;
+import org.example.mollyapi.product.entity.ProductItem;
+import org.example.mollyapi.product.repository.ProductItemRepository;
 import org.example.mollyapi.user.dto.GetUserSummaryInfoWithPointResDto;
 import org.example.mollyapi.user.entity.User;
 import org.example.mollyapi.user.repository.UserRepository;
@@ -50,10 +56,10 @@ import static java.lang.Math.toIntExact;
 @RequiredArgsConstructor
 public class PaymentServiceImpl implements PaymentService {
 
-    private final WebClientUtil webClientUtil;
+    private final DeliveryRepository deliveryRepository;
+    private final ProductItemRepository productItemRepository;
     private final PaymentRepository paymentRepository;
     private final OrderRepository orderRepository;
-    private final OrderService orderService;
     private final PaymentWebClientUtil paymentWebClientUtil;
     private final UserService userService;
     private final UserRepository userRepository;
@@ -82,7 +88,7 @@ public class PaymentServiceImpl implements PaymentService {
 
         // order findByTossOrderId
         Order order = orderRepository.findByTossOrderId(tossOrderId)
-                .orElseThrow(()-> new CustomException(PaymentError.ORDER_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(PaymentError.ORDER_NOT_FOUND));
         Long orderAmount = order.getTotalAmount();
 
         // 유저 포인트 검증
@@ -109,7 +115,7 @@ public class PaymentServiceImpl implements PaymentService {
 
         // 결제 성공 및 실패 로직
         if (res) {
-            successPayment(payment, tossOrderId, point,  deliveryInfoJson);
+            successPayment(payment, tossOrderId, point, deliveryInfoJson);
         } else {
             failPayment(payment, tossOrderId, "실패");
         }
@@ -129,7 +135,7 @@ public class PaymentServiceImpl implements PaymentService {
         TossCancelReqDto tossCancelReqDto = new TossCancelReqDto(paymentCancelReqDto.cancelReason(), paymentCancelReqDto.cancelAmount());
 
         //tossApi 호출
-        ResponseEntity<TossCancelResDto> response = tossPaymentCancelApi(tossCancelReqDto,paymentCancelReqDto.paymentKey());
+        ResponseEntity<TossCancelResDto> response = tossPaymentCancelApi(tossCancelReqDto, paymentCancelReqDto.paymentKey());
 
         // response 정합성 검사
         boolean res = validateResponse(response);
@@ -138,7 +144,9 @@ public class PaymentServiceImpl implements PaymentService {
         TossCancelResDto tossResDto = response.getBody();
 
         // 취소 성공 로직 (payment 상태 canceled 로 변경)
-        if(res){ payment.cancelPayment(); }
+        if (res) {
+            payment.cancelPayment();
+        }
 
         // 성공 여부 리턴
         return res;
@@ -161,7 +169,7 @@ public class PaymentServiceImpl implements PaymentService {
     }
 
     @Override
-    public List<PaymentInfoResDto> findAllPayments(Long orderId){
+    public List<PaymentInfoResDto> findAllPayments(Long orderId) {
         List<Payment> payments = paymentRepository.findAllByOrderByCreatedAtDesc()
                 .orElseThrow(() -> new CustomException(PaymentError.PAYMENT_NOT_FOUND));
         return payments.stream()
@@ -193,7 +201,7 @@ public class PaymentServiceImpl implements PaymentService {
         //payment status change
         payment.successPayment(point);
         //order success (field update, point usage)
-        orderService.successOrder(
+        successOrder(
                 tossOrderId,
                 payment.getPaymentKey(),
                 payment.getPaymentType(),
@@ -208,7 +216,7 @@ public class PaymentServiceImpl implements PaymentService {
      */
     public void failPayment(Payment payment, String tossOrderId, String failureReason) {
         payment.failPayment(failureReason);
-//        orderService.failOrder(tossOrderId);
+        failOrder(tossOrderId);
         throw new CustomException(PaymentError.PAYMENT_FAILED);
     }
 
@@ -218,7 +226,7 @@ public class PaymentServiceImpl implements PaymentService {
      */
     private ResponseEntity<TossConfirmResDto> tossPaymentApi(TossConfirmReqDto tossConfirmReqDto, String apiKey) {
 
-        TossConfirmResDto tossConfirmResDto = paymentWebClientUtil.confirmPayment(tossConfirmReqDto,apiKey);
+        TossConfirmResDto tossConfirmResDto = paymentWebClientUtil.confirmPayment(tossConfirmReqDto, apiKey);
         return ResponseEntity.ok(tossConfirmResDto);
     }
 
@@ -235,9 +243,9 @@ public class PaymentServiceImpl implements PaymentService {
      */
     public Payment createPayment(Long userId, Long orderId, String tossOrderId, String paymentKey, String paymentType, Long amount) {
         User user = userRepository.findById(userId)
-                .orElseThrow(()-> new CustomException(UserError.NOT_EXISTS_USER));
+                .orElseThrow(() -> new CustomException(UserError.NOT_EXISTS_USER));
         Order order = orderRepository.findById(orderId)
-                .orElseThrow(()-> new CustomException(PaymentError.ORDER_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(PaymentError.ORDER_NOT_FOUND));
 
         Payment payment = Payment.from(user, order, tossOrderId, paymentKey, paymentType, amount, "결제대기");
         paymentRepository.save(payment);
@@ -256,7 +264,7 @@ public class PaymentServiceImpl implements PaymentService {
         결제 금액 검증
      */
     private void validateAmount(Long orderAmount, Long amount) {
-        if (!Objects.equals(amount,orderAmount)) {
+        if (!Objects.equals(amount, orderAmount)) {
             throw new CustomException(PaymentError.PAYMENT_AMOUNT_MISMATCH);
         }
     }
@@ -273,4 +281,90 @@ public class PaymentServiceImpl implements PaymentService {
             throw new CustomException(PaymentError.PAYMENT_POINT_INSUFFICIENT);
         }
     }
+
+    public void successOrder(String tossOrderId, String paymentId, String paymentType, Long paymentAmount, Integer pointUsage, String deliveryInfoJson) {
+        // 주문 찾기
+        Order order = orderRepository.findByTossOrderId(tossOrderId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 주문을 찾을 수 없습니다. tossOrderId=" + tossOrderId));
+
+        // 주문 상태 변경
+        order.setStatus(OrderStatus.SUCCEEDED);
+
+        // 사용자의 포인트 차감
+        User user = order.getUser();
+        if (pointUsage != null && pointUsage > 0) {
+            if (user.getPoint() < pointUsage) {
+                throw new IllegalArgumentException("사용자 포인트가 부족합니다.");
+            }
+            user.updatePoint(-pointUsage); // 포인트 차감
+
+            userRepository.save(user);
+        }
+
+        // 결제 정보 업데이트
+        order.updatePaymentInfo(paymentId, paymentType, paymentAmount, pointUsage);
+
+        // 배송 정보 생성
+        createDelivery(order, deliveryInfoJson);
+
+        orderRepository.save(order);
+    }
+
+    private void createDelivery(Order order, String deliveryInfoJson) {
+        try {
+            ObjectMapper objectMapper = new ObjectMapper();
+            JsonNode deliveryInfo = objectMapper.readTree(deliveryInfoJson);
+
+            String receiverName = deliveryInfo.get("receiver_name").asText();
+            String receiverPhone = deliveryInfo.get("receiver_phone").asText();
+            String roadAddress = deliveryInfo.get("road_address").asText();
+            String numberAddress = deliveryInfo.has("number_address") ? deliveryInfo.get("number_address").asText() : null;
+            String addrDetail = deliveryInfo.get("addr_detail").asText();
+
+            // 배송 정보 생성
+            Delivery delivery = Delivery.from(order, receiverName, receiverPhone, roadAddress, numberAddress, addrDetail);
+
+            // 배송 정보 저장
+            deliveryRepository.save(delivery);
+
+            // Order와 연결
+            order.setDelivery(delivery);
+
+            log.info("배송 생성 완료: 주문번호={}, 배송번호={}", order.getId(), delivery.getId());
+
+        } catch (Exception e) {
+            log.error("배송 정보 파싱 실패: {}", e.getMessage());
+            throw new RuntimeException("배송 정보를 저장할 수 없습니다.");
+        }
+    }
+
+    public void failOrder(String tossOrderId) {
+        Order order = orderRepository.findByTossOrderId(tossOrderId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 주문을 찾을 수 없습니다. tossOrderId=" + tossOrderId));
+
+        order.setStatus(OrderStatus.FAILED);
+
+        // 재고 복구
+        for (OrderDetail detail : order.getOrderDetails()) {
+            ProductItem productItem = detail.getProductItem();
+            if (productItem != null) {
+                log.info("[Before] 재고 복구 전 - 상품 ID: {}, 기존 재고: {}, 주문 수량: {}",
+                        productItem.getId(), productItem.getQuantity(), detail.getQuantity());
+
+                productItem.restoreStock(detail.getQuantity()); // 재고 복구
+                productItemRepository.save(productItem);
+                productItemRepository.flush();
+
+                log.info("[After] 재고 복구 완료 - 상품 ID: {}, 실행 후 재고: {}",
+                        productItem.getId(), productItem.getQuantity());
+            } else {
+                log.warn("ProductItem이 null입니다. OrderDetail ID: {}", detail.getId());
+            }
+        }
+
+
+        // 주문 데이터 삭제 (Cascade로 OrderDetail도 삭제됨)
+        orderRepository.delete(order);
+    }
+
 }
